SNOW-1665770: Move groupby.py into src/…/modin/plugin/extensions/
