+++ docs/source/modin/groupby.rst
@@ -2,7 +2,7 @@
 GroupBy
 =============================
 
-.. currentmodule:: snowflake.snowpark.modin.pandas.groupby
+.. currentmodule:: snowflake.snowpark.modin.plugin.extensions.groupby_overrides
 .. rubric:: :doc:`All supported groupby APIs <supported/groupby_supported>`
 
 .. rubric:: Indexing, iteration
