+++ src/snowflake/snowpark/modin/plugin/_internal/telemetry.py
@@ -542,7 +542,7 @@ class TelemetryMeta(type):
     def __new__(
         cls, name: str, bases: tuple, attrs: dict[str, Any]
     ) -> Union[
-        "snowflake.snowpark.modin.pandas.groupby.DataFrameGroupBy",
+        "snowflake.snowpark.modin.plugin.extensions.groupby_overrides.DataFrameGroupBy",
         "snowflake.snowpark.modin.pandas.resample.Resampler",
         "snowflake.snowpark.modin.pandas.window.Window",
         "snowflake.snowpark.modin.pandas.window.Rolling",
@@ -555,7 +555,7 @@ class TelemetryMeta(type):
         with ``snowpark_pandas_telemetry_api_usage`` telemetry decorator.
         Method arguments returned by _get_kwargs_telemetry are collected otherwise set telemetry_args=list().
         TelemetryMeta is only set as the metaclass of:
-         snowflake.snowpark.modin.pandas.groupby.DataFrameGroupBy,
+         snowflake.snowpark.modin.plugin.extensions.groupby_overrides.DataFrameGroupBy,
          snowflake.snowpark.modin.pandas.resample.Resampler,
          snowflake.snowpark.modin.pandas.window.Window,
          snowflake.snowpark.modin.pandas.window.Rolling, and their subclasses.
@@ -567,7 +567,7 @@ class TelemetryMeta(type):
             attrs (Dict[str, Any]): The attributes of the class.
 
         Returns:
-            Union[snowflake.snowpark.modin.pandas.groupby.DataFrameGroupBy,
+            Union[snowflake.snowpark.modin.plugin.extensions.groupby_overrides.DataFrameGroupBy,
                 snowflake.snowpark.modin.pandas.resample.Resampler,
                 snowflake.snowpark.modin.pandas.window.Window,
                 snowflake.snowpark.modin.pandas.window.Rolling]:
